refactor: consolidate hashing support under hash
