+++ datasketches/src/hash/value/raw_bytes.rs
@@ -26,8 +26,8 @@
 
 use std::hash::Hasher;
 
-use super::value::HashStrategy;
-use super::value::Value;
+use super::HashStrategy;
+use super::Value;
 
 /// A byte or string value wrapper that hashes raw bytes.
 ///
@@ -45,8 +45,8 @@ pub struct RawBytesStrategy;
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::raw_bytes::{from_slice, from_vec};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::raw_bytes::{from_slice, from_vec};
 /// assert_eq!(
 ///     calculate_hash(from_vec(b"abc".to_vec())),
 ///     calculate_hash(from_slice(b"abc"))
@@ -65,8 +65,8 @@ pub fn from_vec(v: Vec<u8>) -> RawBytes<Vec<u8>> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::raw_bytes::{from_str, from_string};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::raw_bytes::{from_str, from_string};
 /// assert_eq!(
 ///     calculate_hash(from_string("abc".to_owned())),
 ///     calculate_hash(from_str("abc"))
@@ -85,8 +85,8 @@ pub fn from_string(v: String) -> RawBytes<String> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::raw_bytes::{from_slice, from_vec};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::raw_bytes::{from_slice, from_vec};
 /// assert_eq!(
 ///     calculate_hash(from_slice(b"abc")),
 ///     calculate_hash(from_vec(b"abc".to_vec()))
@@ -105,8 +105,8 @@ pub fn from_slice(v: &[u8]) -> RawBytes<&[u8]> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::raw_bytes::{from_str, from_string};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::raw_bytes::{from_str, from_string};
 /// assert_eq!(
 ///     calculate_hash(from_str("abc")),
 ///     calculate_hash(from_string("abc".to_owned()))
