+++ datasketches/src/hash/value/sign_extend.rs
@@ -25,8 +25,8 @@
 use std::hash::Hash;
 use std::hash::Hasher;
 
-use super::value::HashStrategy;
-use super::value::Value;
+use super::HashStrategy;
+use super::Value;
 
 /// An integer value wrapper that sign-extends the value before hashing.
 ///
@@ -42,8 +42,8 @@ pub struct SignExtendStrategy;
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i8, from_u8};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i8, from_u8};
 /// assert_eq!(calculate_hash(from_i8(-1)), calculate_hash(from_u8(255)));
 /// assert_eq!(calculate_hash(from_i8(-1)), calculate_hash(-1i64));
 /// assert_eq!(calculate_hash(from_i8(42)), calculate_hash(42i64));
@@ -59,8 +59,8 @@ pub fn from_i8(v: i8) -> SignExtend<i8> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i8, from_u8};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i8, from_u8};
 /// assert_eq!(calculate_hash(from_u8(255)), calculate_hash(from_i8(-1)));
 /// assert_eq!(calculate_hash(from_u8(255)), calculate_hash(-1i64));
 /// assert_eq!(calculate_hash(from_u8(1)), calculate_hash(1i64));
@@ -74,8 +74,8 @@ pub fn from_u8(v: u8) -> SignExtend<u8> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i16, from_u16};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i16, from_u16};
 /// assert_eq!(
 ///     calculate_hash(from_i16(-1)),
 ///     calculate_hash(from_u16(65535))
@@ -94,8 +94,8 @@ pub fn from_i16(v: i16) -> SignExtend<i16> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i16, from_u16};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i16, from_u16};
 /// assert_eq!(
 ///     calculate_hash(from_u16(65535)),
 ///     calculate_hash(from_i16(-1))
@@ -112,8 +112,8 @@ pub fn from_u16(v: u16) -> SignExtend<u16> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i32, from_u32};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i32, from_u32};
 /// assert_eq!(
 ///     calculate_hash(from_i32(-1)),
 ///     calculate_hash(from_u32(4294967295))
@@ -132,8 +132,8 @@ pub fn from_i32(v: i32) -> SignExtend<i32> {
 /// # Examples
 ///
 /// ```
-/// # use datasketches::hash_value::calculate_hash;
-/// # use datasketches::hash_value::sign_extend::{from_i32, from_u32};
+/// # use datasketches::hash::value::calculate_hash;
+/// # use datasketches::hash::value::sign_extend::{from_i32, from_u32};
 /// assert_eq!(
 ///     calculate_hash(from_u32(4294967295)),
 ///     calculate_hash(from_i32(-1))
