+++ datasketches/src/hll/sketch.rs
@@ -161,8 +161,8 @@ impl HllSketch {
     /// Accepts any type that implements [`Hash`]. The value is hashed and converted to
     /// an internal coupon, which is then inserted into the sketch.
     ///
-    /// You may use [`hash_value`](crate::hash_value) wrappers when matching other datasketches
-    /// implementations require a specific value hashing strategy.
+    /// You may use [`hash::value`](crate::hash::value) wrappers when another DataSketches
+    /// implementation requires a specific value hashing strategy.
     ///
     /// If you need to insert the same logical value into multiple sketches, consider
     /// pre-computing the coupon with [`Coupon::from_value`] and calling
@@ -174,13 +174,13 @@ impl HllSketch {
     /// ```
     /// # use datasketches::hll::HllSketch;
     /// # use datasketches::hll::HllType;
-    /// # use datasketches::hash_value;
+    /// # use datasketches::hash::value;
     /// let mut sketch = HllSketch::new(10, HllType::Hll8);
     /// sketch.update("apple");
     /// assert!(sketch.estimate() >= 1.0);
     ///
     /// let mut sketch = HllSketch::new(10, HllType::Hll8);
-    /// sketch.update(hash_value::raw_bytes::from_str("apple"));
+    /// sketch.update(value::raw_bytes::from_str("apple"));
     /// assert!(sketch.estimate() >= 1.0);
     /// ```
     pub fn update<T: Hash>(&mut self, value: T) {
